Catches everything up with published SDL1 version La07p70.
I went into aiobject.cpp's "AIObj::newinit()" function and tuned up the flight models
of all of the dive bombers because several of them were underpowered and/or insufficiently
durable.

I replaced main.cpp's "TargetVocalizeProcess1()" with the corresponding (more accurate)
function from SDL1's published version Lac07p70. This required a new global floating-point
variable named "TargetAltitude", initialized near the top of main.cpp. I also enhanced
cockpit.cpp's "drawCounter()" function to populate that variable, cleaning up some of its
code a bit in the process.

I changed the value of main.cpp's "Me163LandingTimer" from 60,000 to 90,000 to force Komet
pilots to wait an additional 30 seconds when refueling.

I went into main.cpp's "event_targetNext()" function and duplicated the logic that was
preventing selection of objects 24, 25, 26, and 27 to additionally prevent selection of
objects 22 and 23. This avoids a segmentation error that can result when those objects are
selected and then subsequently destroyed. Then I made similar changes in main.cpp's very
similar "event_targetPrevious()" function by duplicating the appropriate sections of the
logic it was already using to prevent selection of objects 24, 25, 26, and 27 so that it
cannot be used to select objects 22 or 23. As a consequence of these changes, the
aforementioned segmentation faults no longer occur, but the player can no longer select the
battleship as his actively designated target in either of the island terrain missions. This
is of very little consequence, since he can always select the adjacent airfield and easily
find the battleship thereafter.